move setup_tools functions into compilers.py
