+++ cime_config/cesm/config_files.xml
@@ -40,6 +40,14 @@
     <desc>file containing machine specifications for target model primary component (for documentation only - DO NOT EDIT)</desc>
   </entry>
 
+  <entry id="COMPILERS_SPEC_FILE">
+    <type>char</type>
+    <default_value>$CIMEROOT/cime_config/$MODEL/machines/config_compilers.xml</default_value>
+    <group>case_last</group>
+    <file>env_case.xml</file>
+    <desc>file containing compiler specifications for target model primary component (for documentation only - DO NOT EDIT)</desc>
+  </entry>
+
   <entry id="PIO_SPEC_FILE">
     <type>char</type>
     <default_value>$CIMEROOT/cime_config/$MODEL/machines/config_pio.xml</default_value>
