+++ utils/python/CIME/SystemTests/sms.py
@@ -4,6 +4,7 @@ CIME smoke test  This class inherits from SystemTestsCommon
 from CIME.XML.standard_module_setup import *
 from system_tests_common import SystemTestsCommon
 
+logger = logging.getLogger(__name__)
 
 class SMS(SystemTestsCommon):
     def __init__(self, caseroot, case):
@@ -17,7 +18,11 @@ class SMS(SystemTestsCommon):
         self._case.set_value("REST_OPTION","none")
         self._case.set_value("HIST_OPTION","$STOP_OPTION")
         self._case.set_value("HIST_N","$STOP_N")
+        stop_n = self._case.get_value("STOP_N")
+        stop_option = self._case.get_value("STOP_OPTION")
+
         self._case.flush()
+        logger.info("doing an %d %s initial test, no restarts written"%(stop_n, stop_option))
         SystemTestsCommon.run(self)
 
     def report(self):
