+++ utils/python/CIME/SystemTests/system_tests_common.py
@@ -9,6 +9,8 @@ from CIME.utils import run_cmd
 from CIME.case_setup import case_setup
 import CIME.build as build
 
+logger = logging.getLogger(__name__)
+
 class SystemTestsCommon(object):
     def __init__(self, caseroot=os.getcwd(), case=None):
         """
@@ -21,10 +23,9 @@ class SystemTestsCommon(object):
         # Needed for sh scripts
         os.environ["CASEROOT"] = caseroot
         if case is None:
-            self._case = Case()
+            self._case = Case(caseroot)
         else:
             self._case = case
-
 
         if os.path.isfile(os.path.join(caseroot, "LockedFiles", "env_run.orig.xml")):
             self.compare_env_run()
@@ -38,24 +39,60 @@ class SystemTestsCommon(object):
                         os.path.join(lockedfiles, "env_run.orig.xml"))
 
         self._case.set_initial_test_values()
-        case_setup(self._caseroot, reset=True)
+        case_setup(self._caseroot, reset=True, test_mode=True)
 
 
     def build(self, sharedlib_only=False, model_only=False):
         build.case_build(self._caseroot, case=self._case,
                          sharedlib_only=sharedlib_only, model_only=model_only)
 
+
     def run(self):
-        run_cmd("./case.run")
+        with open("TestStatus", 'a') as f:
+            f.write("PEND %s RUN\n"%self._case.get_value("CASEBASEID"))
+
+        with open("TestStatus", 'r') as f:
+            teststatusfile = f.read()
+
+        rc, out, err = run_cmd("./case.run", ok_to_fail=True)
+        if rc == 0:
+            result = "PASS"
+        else:
+            result = "FAIL"
+
+        with open("TestStatus.log", 'a') as f:
+            f.write("case.run output is %s"%out)
+            f.write("case.run error is %s"%err)
+
+        li = teststatusfile.rsplit('PEND', 1)
+        teststatusfile = result.join(li)
+        with open("TestStatus", 'w') as f:
+            f.write(teststatusfile)
+
+
         return
 
     def report(self):
+        self._checkformemleak()
         return
 
-    def check_mem_leak(self):
-        """ TODO: incomplete """
-        rundir = self._case.get_value("RUNDIR")
-        cpllogfile = min(glob.iglob(os.path.join(rundir, "cpl.log*")), key=os.path.getctime)
+    def _checkformemleak(self):
+        """
+        Examine memory usage as recorded in the cpl log file and look for unexpected
+        increases.
+        """
+
+        newestcpllogfile = self._getlatestcpllog()
+        cmd = os.path.join(self._case.get_value("SCRIPTSROOT"),"Tools","check_memory.pl")
+        rc, out, err = run_cmd("%s -file1 %s -m 1.5"%(cmd, newestcpllogfile),ok_to_fail=True)
+        if rc == 0:
+            with open("TestStatus", "a") as fd:
+                fd.write("PASS %s memleak\n"%(self._case.get_value("CASEBASEID")))
+        else:
+            with open(os.path.join(test_dir, "TestStatus.log"), "a") as fd:
+                fd.write("memleak out: %s\n\nerror: %s"%(out,err))
+            with open(os.path.join(test_dir, "TestStatus"), "a") as fd:
+                fd.write("FAIL %s memleak\n"%(self._case.get_value("CASEBASEID")))
 
     def compare_env_run(self, expected=None):
         f1obj = EnvRun(self._caseroot, "env_run.xml")
@@ -72,6 +109,17 @@ class SystemTestsCommon(object):
                     " value or --reset to reset to original"
                 return False
         return True
+
+    def _getlatestcpllog(self):
+        """
+        find and return the latest cpl log file in the run directory
+        """
+        newestcpllogfile = min(glob.iglob(os.path.join(
+                    self._case.get_value('RUNDIR'),'cpl.log.*')), key=os.path.getctime)
+        return newestcpllogfile
+
+
+
 
 class FakeTest(SystemTestsCommon):
